+++ application/models/FleetInfo.php
@@ -33,8 +33,7 @@ class FleetInfo extends CSV_Model
     // provide form validation rules
     public function rules() {
         $config = array(
-            ['field' => 'id', 'label' => 'ID', 'rules' => 'alpha_numeric_spaces|max_length[64]'],
-            ['field' => 'airid', 'label' => 'Plane ID', 'alpha_numeric_spaces|max_length[64]'],
+            ['field' => 'airid', 'label' => 'Plane ID', 'rules' => 'alpha_numeric_spaces|max_length[64]'],
         );
         return $config;
     }